# [Speed] Add buffer to BinaryReader