--- a/.travis.yml
+++ b/.travis.yml
@@ -1,5 +1,5 @@
 language: swift
-osx_image: xcode9.3
+osx_image: xcode10.1
 script:
   - set -o pipefail && xcodebuild test -scheme CSV-macOS
 after_success: