… .swift-version
… Tests/CSVTests/UnicodeTests.swift