Improved speed of reading from file. (6a18924, 028bd2f)

## Result of the performance test

- OS: Ubuntu 16.04
- Swift: 4.2.1-RELEASE
- Build Configuration: `release`
- CSV file info
    - File size: 13.3MB (15310 records, 55 fields)
    - Encoding: UTF-8 (Including ASCII and Japanese)

| | Before (v2.2.1) | After (this branch) | Delta |
| ---- | ---- | ---- | ---- |
| 1st | 35.313s | 5.811s | -83.5% |
| 2nd | 35.964s | 5.746s | -84.0% |
| 3rd | 35.672s | 5.732s | -83.9% |

## Test code

```swift
// main.swift
import Foundation
import CSV

let stream = InputStream(fileAtPath: "/path/to/file.csv")!
let csv = try! CSVReader(stream: stream, hasHeaderRow: true)
while csv.next() != nil {}
```